requirements.txt
Django==1.11.2
django-datetime-widget==0.9.3
django-guardian==1.4.8
django-simple-captcha==0.5.5
flake8==3.4.1
mccabe==0.6.1
mysqlclient==1.3.10
olefile==0.44
Pillow==4.2.1
pycodestyle==2.3.1
pyflakes==1.5.0
pytz==2017.2
six==1.10.0
uWSGI==2.0.15
xlwt==1.3.0
redis==2.10.6
